# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MFG Telegram Bot API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface SimpleResult {
  ok: boolean;
  message?: string;
}

/**
 * @nullable
 */
export type BotStatusMe = {
  /** @nullable */
  username?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  name?: string | null;
} | null;

export interface CampaignStatus {
  active: boolean;
  total?: number;
  sent?: number;
  failed?: number;
  percent?: number;
  elapsed?: number;
  remain?: number;
}

export interface SmsCampaignStatus {
  active: boolean;
  total?: number;
  sent?: number;
  failed?: number;
  percent?: number;
  elapsed?: number;
  remain?: number;
  provider?: string;
}

export interface BotStatus {
  connected: boolean;
  /** @nullable */
  me?: BotStatusMe;
  uptime: number;
  messages: number;
  aiEnabled: boolean;
  hasGroqKey: boolean;
  hasSmmKey: boolean;
  hasSmsKey: boolean;
  campaign?: CampaignStatus;
  smsCampaign?: SmsCampaignStatus;
}

export interface Contact {
  phone: string;
  name: string;
}

export interface SmsCampaignInput {
  contacts: Contact[];
  message: string;
  provider?: string;
  /** @nullable */
  senderId?: string | null;
}

export interface SmsFlashInput {
  phone: string;
  message: string;
  provider?: string;
  /** @nullable */
  senderId?: string | null;
}

export interface SmsFlashResult {
  ok: boolean;
  /** @nullable */
  messageId?: string | null;
  message: string;
  provider: string;
}

export interface SmsProvider {
  id: string;
  name: string;
  configured: boolean;
  /** @nullable */
  balance?: string | null;
}

export interface SmsProviderList {
  providers: SmsProvider[];
  active: string;
}

export interface SmsHistoryItem {
  id: string;
  phone: string;
  message: string;
  status: string;
  provider: string;
  at: number;
  /** @nullable */
  messageId?: string | null;
}

export interface SmsHistoryList {
  items: SmsHistoryItem[];
}

export interface CampaignStartInput {
  contacts: Contact[];
  message: string;
}

export interface LoginStartInput {
  phone: string;
}

export interface LoginCodeInput {
  code: string;
}

export interface Login2faInput {
  password: string;
}

export interface BotSettings {
  aiEnabled: boolean;
  aiMode?: string;
  aiDelay?: number;
  transcribeVoice?: boolean;
  visionEnabled?: boolean;
  antiScam?: boolean;
  moodAware?: boolean;
  birthdayWishes?: boolean;
  autoTakeover?: boolean;
  takeoverMinutes?: number;
  smmMarkup?: number;
  smmNGNRate?: number;
  prefix: string;
  botName: string;
  aiDisclaimer?: boolean;
  disclaimerText?: string;
  systemPrompt?: string;
  smsProvider?: string;
  /** @nullable */
  smsApiKey?: string | null;
  /** @nullable */
  smsSenderId?: string | null;
}

export interface BotSettingsUpdate {
  aiEnabled?: boolean;
  aiMode?: string;
  aiDelay?: number;
  transcribeVoice?: boolean;
  visionEnabled?: boolean;
  antiScam?: boolean;
  moodAware?: boolean;
  birthdayWishes?: boolean;
  autoTakeover?: boolean;
  takeoverMinutes?: number;
  smmMarkup?: number;
  smmNGNRate?: number;
  prefix?: string;
  botName?: string;
  aiDisclaimer?: boolean;
  disclaimerText?: string;
  systemPrompt?: string;
  smsProvider?: string;
  /** @nullable */
  smsApiKey?: string | null;
  /** @nullable */
  smsSenderId?: string | null;
}

export type WalletTopupsItem = { [key: string]: unknown };

export type WalletSpendsItem = { [key: string]: unknown };

export interface Wallet {
  userId: string;
  balance: number;
  currency: string;
  topups?: WalletTopupsItem[];
  spends?: WalletSpendsItem[];
}

export interface WalletTopupInput {
  userId: string;
  amount: number;
  note?: string;
}

export interface ScamAlert {
  chatId: string;
  text: string;
  at: number;
}

export interface ScamAlertList {
  alerts: ScamAlert[];
}