# lib/api-zod/src/generated/types/campaignStartInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MFG Telegram Bot API
 * OpenAPI spec version: 0.1.0
 */
import type { Contact } from './contact';

export interface CampaignStartInput {
  contacts: Contact[];
  message: string;
}