# lib/api-zod/src/generated/types/smsCampaignStatus.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MFG Telegram Bot API
 * OpenAPI spec version: 0.1.0
 */

export interface SmsCampaignStatus {
  active: boolean;
  total?: number;
  sent?: number;
  failed?: number;
  percent?: number;
  elapsed?: number;
  remain?: number;
  provider?: string;
}