# lib/api-zod/src/generated/types/smsFlashInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MFG Telegram Bot API
 * OpenAPI spec version: 0.1.0
 */

export interface SmsFlashInput {
  phone: string;
  message: string;
  provider?: string;
  /** @nullable */
  senderId?: string | null;
}