# lib/api-zod/src/generated/types/smsHistoryList.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MFG Telegram Bot API
 * OpenAPI spec version: 0.1.0
 */
import type { SmsHistoryItem } from './smsHistoryItem';

export interface SmsHistoryList {
  items: SmsHistoryItem[];
}