# lib/api-zod/src/generated/types/walletTopupInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MFG Telegram Bot API
 * OpenAPI spec version: 0.1.0
 */

export interface WalletTopupInput {
  userId: string;
  amount: number;
  note?: string;
}